Add Context object for WebSocket callbacks
`on_open`, `on_message`, and `on_error` callbacks for `WebSocketClient`
now take an additional `Context` object as an argument. `Context`
contains a copy of the original `WebSocketClient` and it's `Igniter`. This is
meant to make it easier to restart a WebSocket connection, start another
WebSocket,  and dispatch http futures from within a WebSocket callback.

Signed-off-by: Ryan Banks <rbanks@bitwise.io>